refactor(watch): split static seam fields + Unity-boundary Safe wrappers into WatchModeController.Runtime.cs
